card-dsl: move Skills + SkillKind down from game-core
Relocate the pure-data Skills struct (+ its value() method) and SkillKind
enum into card_dsl::card_data, alongside the existing skill vocabulary
(Stat, SkillTestKind). game_core::state re-exports both at their historical
paths, so all references compile unchanged. Prep for CardKind::Investigator
to carry skills: Skills (a card-dsl type) in the upcoming remodel.
