+++ crates/game-core/src/state/game_state.rs
@@ -8,11 +8,12 @@ use super::{
     card::{CardCode, CardInstanceId},
     chaos_bag::{ChaosBag, TokenModifiers},
     enemy::{Enemy, EnemyId},
-    investigator::{Investigator, InvestigatorId, SkillKind},
+    investigator::{Investigator, InvestigatorId},
     location::{Location, LocationId},
     phase::Phase,
 };
 use crate::dsl::{SkillTestKind, Stat};
+use card_dsl::card_data::SkillKind;
 use crate::rng::RngState;
 
 /// The full state of a scenario at a single point in time.
